Typography tokens
BREAKING CHANGE: This changes the types for color tokens.
