+++ packages/design-tokens/tokens/typography.json
@@ -1,0 +1,57 @@
+{
+  "kz": {
+    "typography": {
+      "family": {
+        "mono": "\"IBM Plex Mono\"",
+        "display": "\"Greycliff CF\"",
+        "text": "\"Inter\""
+      },
+      "weight": {
+        "bold": "600",
+        "regular": "400"
+      },
+      "display": {
+        "fontSize": {
+          "1": "72px",
+          "2": "36px",
+          "3": "30px",
+          "4": "22px",
+          "5": "18px",
+          "6": "16px",
+          "7": "14px"
+        },
+        "lineHeight": {
+          "1": "72px",
+          "2": "36px",
+          "3": "36px",
+          "4": "24px",
+          "5": "24px",
+          "6": "24px",
+          "7": "24px"
+        }
+      },
+      "text": {
+        "fontSize": {
+          "1": "20px",
+          "2": "16px",
+          "3": "14px",
+          "4": "12px"
+        },
+        "lineHeight": {
+          "1": "30px",
+          "2": "24px",
+          "3": "18px",
+          "4": "18px"
+        }
+      },
+      "mono": {
+        "fontSize": {
+          "1": "16px"
+        },
+        "lineHeight": {
+          "1": "24px"
+        }
+      }
+    }
+  }
+}
